chg: refactor package structure and introduce Fetch for extending fetchers
Refactor:
+ Move Colusa class to separate module colusa, previously within __init__.py
+ Move download_url and download_image from utils to fetch

New Extension Point:
+ Introduce Fetch class which replicate the requests methods and allow inheritance
so that plugins can implement extended functionalities if needed
