+++ src/colusa/crawlers.py
@@ -3,14 +3,14 @@ from bs4 import BeautifulSoup
 import pathlib
 from collections import OrderedDict
 from urllib.parse import urljoin
-from .utils import scan, download_url, get_hexdigest
-
+from colusa import utils
 
 class Crawler(object):
     def __init__(self, url, output_dir, output_file):
         self.url = url
         self.output_dir = output_dir
         self.output_file = output_file
+        self.downloader = utils.Downloader()
 
     def run(self):
         logs.info(f"hello: {self.url}")
@@ -37,12 +37,12 @@ class Crawler(object):
         :return: content of downloaded file
         """
         output_path = pathlib.Path(self.output_dir)
-        cached_file_path = output_path.joinpath('.cached', f'{get_hexdigest(self.url)}.html')
+        cached_file_path = output_path.joinpath('.cached', f'{utils.get_hexdigest(self.url)}.html')
         logs.info(self.url, cached_file_path)
 
         if not cached_file_path.exists():
             # download file from url_path
-            download_url(self.url, str(cached_file_path))
+            self.downloader.download_url(self.url, str(cached_file_path))
 
         with open(cached_file_path, 'rt', encoding='utf-8') as file_in:
             content = file_in.read()
