+++ src/colusa/plugins/pp_regex.py
@@ -1,7 +1,6 @@
 from colusa.etr import PostProcessor, register_postprocessor
 import re
 
-
 @register_postprocessor('post-processor.RegexReplace')
 class RegexReplaceProcessor(PostProcessor):
     def run(self):
@@ -9,7 +8,6 @@ class RegexReplaceProcessor(PostProcessor):
             data = fd.read()
 
         for kv in self.params:
-            print(kv.get('s'))
             data = re.sub(kv.get('s'), kv.get('r'), data, flags=re.MULTILINE)
 
         with open(self.file_path, 'wt') as fd:
